readers-haven/requirements.txt
Django==2.2.3
django-braces==1.13.0
django-materializecss-form==1.1.14
pytz==2019.1
sqlparse==0.3.0
